chore(ci)!: partially drop eslint, prettier in favor of biome
