+++ biome.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://biomejs.dev/schemas/1.7.2/schema.json",
+  "extends": ["packages/biome-config/biome.jsonc"]
+}
