+++ package.json
@@ -3,21 +3,27 @@
   "private": true,
   "scripts": {
     "prepare": "is-ci || husky install .github/husky",
-    "lint": "eslint packages --ext mjs,js,ts --fix",
-    "format": "prettier --write \"packages/**/{src,tests,scripts}/**/*.{mjs,js,ts}\"",
+    "dry-format": "biome check --no-errors-on-unmatched --files-ignore-unknown=true",
+    "format": "biome check --no-errors-on-unmatched --files-ignore-unknown=true --apply",
     "test": "vitest run",
     "build": "turbo run build",
     "bump": "turbo run bump",
     "check-update": "turbo run check-update"
   },
+  "lint-staged": {
+    "*": [
+      "pnpm format -- --staged"
+    ]
+  },
   "devDependencies": {
+    "@biomejs/biome": "1.7.3",
     "@commitlint/cli": "^18.6.1",
     "@commitlint/config-angular": "^18.6.1",
     "@favware/cliff-jumper": "^2.2.3",
     "@types/node": "^20.12.8",
     "@typescript-eslint/eslint-plugin": "^7.8.0",
     "@typescript-eslint/parser": "^7.8.0",
-    "@vitest/coverage-c8": "^0.33.0",
+    "@vitest/coverage-v8": "^1.6.0",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
