+++ packages/eslint-config/tsup.config.ts
@@ -1,3 +1,4 @@
-import { createTsupConfig } from '../../scripts/tsup.config'
+import { createTsupConfig } from "../../scripts/tsup.config";
 
-export default createTsupConfig({ entry: ['src/*.ts'], sourcemap: false })
+// biome-ignore lint/style/noDefaultExport: This is a config file
+export default createTsupConfig({ entry: ["src/*.ts"], sourcemap: false });
