--- packages/eslint-config/vitest.config.ts
@@ -1,3 +1,0 @@
-import { createVitestConfig } from '../../scripts/vitest.config'
-
-export default createVitestConfig
