+++ packages/prettier-config/vitest.config.mts
@@ -1,0 +1,4 @@
+import { createVitestConfig } from "../../scripts/vitest.config.mjs";
+
+// biome-ignore lint/style/noDefaultExport: This is a config file
+export default createVitestConfig;
